pyproject.toml
[project]
name = "solana-agent-api"
version = "0.0.1"
description = ""
authors = [{ name = "Bevan Hunt", email = "bevan@bevanhunt.com" }]
requires-python = ">=3.13.11"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "gunicorn>=23.0.0",
    "python-dotenv>=1",
    "solana-agent>=33.1.1",
    "sakit>=17.2.2",
    "python-jose[cryptography]>=3.5.0",
    "websockets",
    "telethon==1.42.0",
    "pillow==12.1.0",
    "cryptg==0.5.2",
    "aiohttp==3.13.3",
    "motor>=3.7.1",
    "pyjwt>=2.10.1",
    "httpx>=0.28.0",
    "solders>=0.27.1",
    "segno>=1.6.6",
    "nanoid>=2.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.1.0",
    "pytest-mock>=3.14.0",
    "mongomock-motor>=0.0.34",
    "respx>=0.22.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["solana_agent_api"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["solana_agent_api"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
